requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pycparser==2.21
PyJWT==1.7.1
requests==2.28.1
six==1.16.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.19.9
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
